feat: better routing, documentation
