+++ src/app/page.js
@@ -1,101 +1,81 @@
-import Image from "next/image";
+'use client';
 
-export default function Home() {
-  return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-        <Image
-          className="dark"
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol className="list-inside list-decimal text-sm text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
-          <li className="mb-2">
-            Get started by editing{" "}
-            <code className="bg-black/[.05] dark:bg-white/[.06] px-1 py-0.5 rounded font-semibold">
-              src/app/page.js
-            </code>
-            .
-          </li>
-          <li>Save and see your changes instantly.</li>
-        </ol>
+import React, { useState } from "react"
 
-        <div className="flex gap-4 items-center flex-col sm:flex-row">
-          <a
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5"
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className="dark"
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:min-w-44"
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Read our docs
-          </a>
+export default function Page1() {
+    // TODO: connect to the server to verify that the game code exists before allowing the user to join
+    // TODO: allow the user to actually join/create the game: send them to the game page and do the necessary updates on the server (ex. setting them as the host)
+    return (
+        <div className="w-full text-center">
+            <p className="text-xl mt-20">
+                Ready for an icebreaker?
+            </p>
+            <br/>
+            <div style={{"display": "flex", "justifyContent": "center"}}>
+                <div style={{"width": "40%"}}>
+                    <JoinBox/>
+                    <br/>
+                    OR
+                    <br/><br/>
+                    <CreateBox/>
+                </div>
+            </div>
+            <p className="m-40">
+                <b>Credits</b>
+                <br/>This was created for the ICSSC WebJam 2024 at UCI by Caden Lee, Hodaka Nakamura, Kelvin Wu, and Kyle Kim.
+                <br/>We used Next.js, React, and other technologies.
+                <br/>Check out the source code and get more information on <a className="link-underline" href="https://github.com/hodaknak/webjam2024">GitHub</a>!
+            </p>
         </div>
-      </main>
-      <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
-    </div>
-  );
+    )
+}
+
+function JoinBox() {
+    const [currentCode, setCurrentCode] = useState("");
+    const handleChange = (event) => {
+        const value = event.target.value;
+        setCurrentCode(value)
+    }
+    const isCodeValid = (code) => {
+        // Depending on current code
+        // TODO: fix this to match however we choose code lengths/validity, and also verify the game exists on the server
+        let isValid = code.length == 4;
+        return isValid;
+    }
+    const keyDown = (event) => {
+        const key = event.key;
+        if (key == "Enter" && isCodeValid(currentCode)) {
+            submit();
+        }
+    }
+    const submit = () => {
+        if (!isCodeValid(currentCode)) {
+            // Invalid code
+            return;
+        }
+        // Try to submit the join request with the code
+        console.log("Submitting the join request with code `" + currentCode + "`...")
+        // TODO: impl
+    }
+    return (
+        <div className="roundbox">
+            <input onChange={handleChange} autoFocus={true} type="text" placeholder="ENTER JOIN CODE" className="codebox" onKeyDown={keyDown}/>
+            <br/>
+            <br/>
+            <button disabled={!isCodeValid(currentCode)} onClick={submit}>Join game</button>
+        </div>
+    )
+}
+
+function CreateBox() {
+    const submit = () => {
+        // Try to submit the create request
+        console.log("Submitting the create request...")
+        // TODO: impl
+    }
+    return (
+        <div className="roundbox">
+            <button onClick={submit}>Create a game</button>
+        </div>
+    )
 }
